+++ packages/react-widgets/src/widgets/LegendWidget.js
@@ -8,10 +8,10 @@ import PropTypes from 'prop-types';
  * Renders a <LegendWidget /> component
  * @param  {object} props
  * @param  {string} props.className - CSS class name
- * @param  {Object.<string, function>} props.legendTypes - Allow to customise by default legend types that can be rendered
+ * @param  {Object.<string, function>} props.customLegendTypes - Allow to customise by default legend types that can be rendered
  * @param  {string} props.initialCollapsed - Define initial collapsed value. false by default.
  */
-function LegendWidget({ className, legendTypes, initialCollapsed }) {
+function LegendWidget({ className, customLegendTypes, initialCollapsed }) {
   const dispatch = useDispatch();
   const layers = useSelector((state) =>
     Object.values(state.carto.layers).filter((layer) => !!layer.legend)
@@ -31,12 +31,22 @@ function LegendWidget({ className, legendTypes, initialCollapsed }) {
     );
   };
 
+  const handleChangeOpacity = ({ id, opacity }) => {
+    dispatch(
+      updateLayer({
+        id,
+        layerAttributes: { opacity }
+      })
+    );
+  };
+
   return (
     <LegendWidgetUI
       className={className}
-      legendTypes={legendTypes}
+      customLegendTypes={customLegendTypes}
       layers={layers}
       onChangeVisibility={handleChangeVisibility}
+      onChangeOpacity={handleChangeOpacity}
       collapsed={collapsed}
       onCollapsedChange={setCollapsed}
     />
@@ -45,7 +55,7 @@ function LegendWidget({ className, legendTypes, initialCollapsed }) {
 
 LegendWidget.propTypes = {
   className: PropTypes.string,
-  legendTypes: PropTypes.objectOf(PropTypes.func),
+  customLegendTypes: PropTypes.objectOf(PropTypes.func),
   initialCollapsed: PropTypes.bool
 };