+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Not released
 
-- Add legendTypes prop to customise covered legend types [#325](https://github.com/CartoDB/carto-react/pull/325)
+- Add customLegendTypes prop to customise covered legend types [#325](https://github.com/CartoDB/carto-react/pull/325)
 - Fix Switch input width [#323](https://github.com/CartoDB/carto-react/pull/323)
+- Add optional slider to modify layer opacity in Legend widget [#324](https://github.com/CartoDB/carto-react/pull/324)
 
 ## 1.2